porting acl functions to new acl class.
