+++ app/controllers/api/v1/servers/status_controller.rb
@@ -1,30 +1,30 @@
 class Api::V1::Servers::StatusController < ApplicationController
-	require 'ipaddr'
-	require 'yaml'
+	# require 'ipaddr'
+	# require 'yaml'
 
-	class DeniedHostError < StandardError
-		def initialize msg="Request to this host is not allowed."
-			super msg
-		end
-	end
+	# class DeniedHostError < StandardError
+	# 	def initialize msg="Request to this host is not allowed."
+	# 		super msg
+	# 	end
+	# end
 
-	DENY_IP_ADDRESSES = [
-		# IPv4
-		"192.168.0.0/16",
-		"172.16.0.0/12",
-		"10.0.0.0/8",
-		"255.255.255.255",   # broadcast
-		"127.0.0.0/8",   # loop back
-		"169.254.0.0/16",   # link local
-		"100.64.0.0/10",   # shared address space ref: RFC6890
+	# DENY_IP_ADDRESSES = [
+	# 	# IPv4
+	# 	"192.168.0.0/16",
+	# 	"172.16.0.0/12",
+	# 	"10.0.0.0/8",
+	# 	"255.255.255.255",   # broadcast
+	# 	"127.0.0.0/8",   # loop back
+	# 	"169.254.0.0/16",   # link local
+	# 	"100.64.0.0/10",   # shared address space ref: RFC6890
 
-		# IPv6
-		"::1/128",   # loop back
-		"fc00::/7",   # unique local
-		"fe80::/10",   # link local
-		"ff01::1",   # multicast on interface local
-		"ff02::1",   # multicast on link local
-	].freeze
+	# 	# IPv6
+	# 	"::1/128",   # loop back
+	# 	"fc00::/7",   # unique local
+	# 	"fe80::/10",   # link local
+	# 	"ff01::1",   # multicast on interface local
+	# 	"ff02::1",   # multicast on link local
+	# ].freeze
 
 	def index
 		@host = params[:host]
@@ -32,9 +32,10 @@ class Api::V1::Servers::StatusController < ApplicationController
 		@port = params[:port].to_i unless params[:port].nil?
 
 		begin
-			host_filter @host
+			@acl = Acl::Acl.new @host
+			@acl.filter!
 			@server = Minetools::ServerStatusTool::ServerStatus.new host: @host, port: @port
-		rescue DeniedHostError => e
+		rescue Acl::DeniedHostError => e
 			render status: 400, json: {message: e.message}
 			return
 		rescue ArgumentError => e
@@ -71,36 +72,36 @@ class Api::V1::Servers::StatusController < ApplicationController
 		}
 	end
 
-	private	def host_filter _host
-		raise ArgumentError unless _host.class == String
-		# TODO: fastly return with white list
+	# private	def host_filter _host
+	# 	raise ArgumentError unless _host.class == String
+	# 	# TODO: fastly return with white list
 
-		is_hostname = _host_name_filter_default _host
-		is_ipaddr = _ip_address_filter_default _host
+	# 	is_hostname = _host_name_filter_default _host
+	# 	is_ipaddr = _ip_address_filter_default _host
 
-		raise ArgumentError, "given host is not IP address or correct hostname." if !is_ipaddr && !is_hostname
-	end
+	# 	raise ArgumentError, "given host is not IP address or correct hostname." if !is_ipaddr && !is_hostname
+	# end
 
-	private def _host_name_filter_default _host
-		_match = /^(?:[a-zA-Z0-9][a-zA-Z0-9-]{,63}[a-zA-Z0-9]{,63}\.){,253}([a-zA-Z]{2,})$/.match _host.upcase
-		return false if _match.nil?
+	# private def _host_name_filter_default _host
+	# 	_match = /^(?:[a-zA-Z0-9][a-zA-Z0-9-]{,63}[a-zA-Z0-9]{,63}\.){,253}([a-zA-Z]{2,})$/.match _host.upcase
+	# 	return false if _match.nil?
 
-		tld_list = App::Application.config.tld_list["TLD"]
-		unless tld_list.map{|i| i.to_s.upcase}.include? _match[1]
-			raise ArgumentError, "given TLD of hostname is not correct."
-		end
-		return true
-	end
+	# 	tld_list = App::Application.config.tld_list["TLD"]
+	# 	unless tld_list.map{|i| i.to_s.upcase}.include? _match[1]
+	# 		raise ArgumentError, "given TLD of hostname is not correct."
+	# 	end
+	# 	return true
+	# end
 
-	private def _ip_address_filter_default _host
-		begin
-			_host_address = IPAddr.new(_host)
-		rescue IPAddr::InvalidAddressError => e
-			return false
-		end
-		DENY_IP_ADDRESSES.each do |addr|
-			raise DeniedHostError if IPAddr.new(addr).include? _host_address
-		end
-		return true
-	end
+	# private def _ip_address_filter_default _host
+	# 	begin
+	# 		_host_address = IPAddr.new(_host)
+	# 	rescue IPAddr::InvalidAddressError => e
+	# 		return false
+	# 	end
+	# 	DENY_IP_ADDRESSES.each do |addr|
+	# 		raise DeniedHostError if IPAddr.new(addr).include? _host_address
+	# 	end
+	# 	return true
+	# end
 end
