+++ spec/lib/acl/acl_spec.rb
@@ -3,7 +3,7 @@ require './lib/acl/acl.rb'
 
 RSpec.describe Acl::Acl do
 	it "is object of Acl::Acl class" do
-		acl = Acl::Acl.new
+		acl = Acl::Acl.new ""
 		expect(acl.class).to eq(Acl::Acl)
 	end
 end
